pyserial/requirements.txt
#pip install -r requirements.txt
flask
pyserial
flask-socketio
threading